+++ client/src/router.js
@@ -2,11 +2,15 @@ const Vue = require('vue')
 const VueRouter = require('vue-router')
 
 const Sheet = require('./pages/sheet.vue')
+const LoginCallback = require('./pages/login-callback.vue')
+const { initiateLogin } = require('./helpers/auth0')
 
 Vue.use(VueRouter)
 
 const routes = [
   { path: '/', component: Sheet },
+  { path: '/login', beforeEnter: initiateLogin },
+  { path: '/login/callback', component: LoginCallback },
   { path: '/:sheetName', component: Sheet }
 ]
