+++ client/src/store/modules/auth.js
@@ -1,0 +1,13 @@
+const api = require('../../api')
+
+module.exports = {
+  state: {
+    isAuthenticated: false
+  },
+  actions: {
+    async finishLogin ({ commit }, authCode) {
+      const user = await api.authenticate(authCode)
+      console.log(user)
+    }
+  }
+}
