+++ package.json
@@ -35,6 +35,7 @@
     "passport-auth0": "^0.6.0",
     "pg": "^6.2.4",
     "postinstall-build-yarn": "^0.0.2",
+    "query-string": "^4.3.4",
     "shortid": "^2.2.8",
     "url-join": "^2.0.2",
     "vue": "^2.0.1",
